server/requirements.txt
bcrypt==4.3.0
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
python-dotenv==1.0.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
